refactor(button): convertLegacyProps function moved to buttonHelpers
