+++ components/popconfirm/PurePanel.tsx
@@ -3,7 +3,7 @@ import ExclamationCircleFilled from '@ant-design/icons/ExclamationCircleFilled';
 import classNames from 'classnames';
 import type { PopconfirmProps } from '.';
 import Button from '../button';
-import { convertLegacyProps } from '../button/button';
+import { convertLegacyProps } from '../button/buttonHelpers';
 import ActionButton from '../_util/ActionButton';
 import LocaleReceiver from '../locale/LocaleReceiver';
 import defaultLocale from '../locale/en_US';
